[level 1] Title: 콜라츠 추측, Time: 0.04 ms, Memory: 76.2 MB -BaekjoonHub

+++ 프로그래머스/1/12943. 콜라츠 추측/콜라츠 추측.java
@@ -1,5 +1,5 @@
 class Solution {
-    public long solution(long num) {
+    public int solution(long num) {
         int answer = 0;
         if(num == 1){
             return 0;
